refactor: Migrate pkg/command to internal/command since we use pkg as a independent concern
