+++ cmd/code/scan.go
@@ -2,11 +2,10 @@ package code
 
 import (
 	"context"
-	"fmt"
 
+	"github.com/safedep/vet/internal/command"
 	"github.com/safedep/vet/internal/ui"
 	"github.com/safedep/vet/pkg/code"
-	"github.com/safedep/vet/pkg/command"
 	"github.com/safedep/vet/pkg/common/logger"
 	"github.com/safedep/vet/pkg/storage"
 	"github.com/spf13/cobra"
@@ -72,7 +71,7 @@ func internalStartScan() error {
 			},
 			OnScanEnd: func() error {
 				ui.StopSpinner()
-				fmt.Println("Scan complete")
+				ui.PrintSuccess("Code scanning completed")
 				return nil
 			},
 		},
